1936: Added abstract client base class with share functions
